Move realm code into identity module and keep authn library as importable APIs
