+++ sandbox/modules/identity/src/main/java/org/opensearch/identity/User.java
@@ -1,11 +1,15 @@
 /*
- * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
  */
 
-package org.opensearch.authn;
+package org.opensearch.identity;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import org.opensearch.authn.StringPrincipal;
 
 import java.util.Collections;
 import java.util.Map;
