+++ sandbox/modules/identity/src/main/resources/shiro.ini
@@ -6,7 +6,7 @@
 # Objects and their properties are defined here,
 # Such as the securityManager, Realms and anything
 # else needed to build the SecurityManager
-myRealm = org.opensearch.authn.realm.InternalRealm
+myRealm = org.opensearch.identity.realm.InternalRealm
 myRealm.realmName = myRealm
 myRealm.internalUsersYaml = example/example_internal_users.yml
